@@ -6,9 +6,11 @@ import com.societegenerale.githubcrawler.model.Repository
 import com.societegenerale.githubcrawler.output.GitHubCrawlerOutput
 import com.societegenerale.githubcrawler.parsers.SimpleFilePathParser
 import com.societegenerale.githubcrawler.remote.RemoteGitHub
+
 import org.assertj.core.api.Assertions.assertThat
-import org.junit.Before
-import org.junit.Test
+import org.junit.jupiter.api.BeforeEach
+import org.junit.jupiter.api.Test
+
 import org.mockito.Mockito
 import org.mockito.Mockito.*
 import org.springframework.core.env.Environment
@@ -36,7 +38,7 @@ class GitHubCrawlerTest {
 
     private lateinit var gitHubCrawler: GitHubCrawler
 
-    @Before
+    @BeforeEach
     fun setUp() {
 
         gitHubCrawler = GitHubCrawler(mockRemoteGitHub, outputs, repositoryEnricher, gitHubCrawlerProperties, mockEnvironment, organizationName, mockConfigValidator,fileContentParsers)
