@@ -5,7 +5,7 @@ import com.societegenerale.githubcrawler.model.Repository
 import com.societegenerale.githubcrawler.remote.RemoteGitHub
 import com.societegenerale.githubcrawler.repoTaskToPerform.RepoTaskToPerform
 import org.assertj.core.api.Assertions.assertThat
-import org.junit.Test
+import org.junit.jupiter.api.Test
 
 import org.mockito.Mockito
 import java.util.*
