@@ -1,7 +1,7 @@
 package com.societegenerale.githubcrawler.model.team
 
 import org.assertj.core.api.Assertions
-import org.junit.Test
+import org.junit.jupiter.api.Test
 
 class MembershipTest {
 
