@@ -2,11 +2,12 @@ package com.societegenerale.githubcrawler.output
 
 import ch.qos.logback.classic.spi.ILoggingEvent
 import ch.qos.logback.core.Appender
-import com.nhaarman.mockito_kotlin.*
 import com.societegenerale.githubcrawler.model.Branch
 import com.societegenerale.githubcrawler.model.Repository
 import org.assertj.core.api.Assertions.assertThat
-import org.junit.Test
+import org.junit.jupiter.api.Test
+import org.mockito.ArgumentCaptor
+import org.mockito.Mockito.*
 import org.slf4j.LoggerFactory
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
@@ -18,87 +19,90 @@ import java.util.*
 
 class HttpOutputTest {
 
-    private val logger = LoggerFactory.getLogger(HttpOutput::class.java)
+  private val logger = LoggerFactory.getLogger(HttpOutput::class.java)
 
-    private val dummyRepo = Repository(name = "repo1",
-            creationDate = Date(),
-            config = null,
-            defaultBranch = "master",
-            fullName = "orgName/repoName1",
-            lastUpdateDate = Date(),
-            reason = null,
-            url = "http://hello",
-            branchesToParse = setOf(Branch("master"))
+  private val dummyRepo = Repository(name = "repo1",
+      creationDate = Date(),
+      config = null,
+      defaultBranch = "master",
+      fullName = "orgName/repoName1",
+      lastUpdateDate = Date(),
+      reason = null,
+      url = "http://hello",
+      branchesToParse = setOf(Branch("master"))
+  )
+
+  private val mockRestTemplate = mock(RestTemplate::class.java)
+
+  inline fun <reified T : Any> argumentCaptor() = ArgumentCaptor.forClass(T::class.java)
+
+  @Test
+  fun shouldLogResponseBodyWhenErrorDuringPost() {
+    //Mock logging infra
+    val root = LoggerFactory.getLogger(ch.qos.logback.classic.Logger.ROOT_LOGGER_NAME) as ch.qos.logback.classic.Logger
+    val mockAppender = mock(Appender::class.java)
+
+    `when`(mockAppender.name).thenReturn("MOCK")
+    root.addAppender(mockAppender as Appender<ILoggingEvent>?)
+
+    val httpOutput = HttpOutput("http://localhost", mockRestTemplate)
+
+    `when`(mockRestTemplate.postForEntity(any<String>(),
+        any<OutputIndicator>(),
+        eq(String::class.java)
     )
+    ).thenThrow(HttpClientErrorException(HttpStatus.BAD_REQUEST, "Bad Request details"))
 
-    private val mockRestTemplate: RestTemplate = mock()
+    //ACTION
+    httpOutput.output(dummyRepo)
 
-    @Test
-    fun shouldLogResponseBodyWhenErrorDuringPost() {
-        //Mock logging infra
-        val root = LoggerFactory.getLogger(ch.qos.logback.classic.Logger.ROOT_LOGGER_NAME) as ch.qos.logback.classic.Logger
-        val mockAppender: Appender<ILoggingEvent> = mock()
+    val captor = argumentCaptor<ILoggingEvent>()
 
-        whenever(mockAppender.name).thenReturn("MOCK")
-        root.addAppender(mockAppender)
+    verify(mockAppender).doAppend(captor.capture())
 
-        val httpOutput = HttpOutput("http://localhost", mockRestTemplate)
+    val matchingLogEvent = captor.allValues.firstOrNull { logEvent -> logEvent.formattedMessage.contains("Bad Request details") }
 
-        whenever(mockRestTemplate.postForEntity(any<String>(),
-                                                any<OutputIndicator>(),
-                                                eq(String::class.java)
-                                             )
-              ).thenThrow(HttpClientErrorException(HttpStatus.BAD_REQUEST, "Bad Request details"))
+    assertThat(matchingLogEvent).`as`("can't find a log statement with expected message").isNotNull
 
-        //ACTION
-        httpOutput.output(dummyRepo)
+  }
 
-        argumentCaptor<ILoggingEvent>().apply {
-            verify(mockAppender).doAppend(capture())
+  @Test
+  fun shouldHitProperUrlWhenPlaceHolderProvided() {
 
-            val matchingLogEvent = allValues.firstOrNull{ logEvent -> logEvent.formattedMessage.contains("Bad Request details") }
+    val httpOutput = HttpOutput("http://localhost/{YYYY-MM}", mockRestTemplate, LocalDate.of(2019, 1, 25))
 
-            assertThat(matchingLogEvent ).`as`("can't find a log statement with expected message").isNotNull
-        }
-    }
+    `when`(mockRestTemplate.postForEntity(any<String>(),
+        any<OutputIndicator>(),
+        eq(String::class.java))
+    ).thenReturn(ResponseEntity("content", HttpStatus.ACCEPTED))
 
-    @Test
-    fun shouldHitProperUrlWhenPlaceHolderProvided() {
+    //ACTION
+    httpOutput.output(dummyRepo)
 
-        val httpOutput = HttpOutput("http://localhost/{YYYY-MM}", mockRestTemplate, LocalDate.of(2019, 1, 25))
+    verify(mockRestTemplate, times(1)).postForEntity(eq("http://localhost/2019-01"),
+        any<OutputIndicator>(),
+        eq(String::class.java)
+    )
+  }
 
-        whenever(mockRestTemplate.postForEntity(any<String>(),
-                                                any<OutputIndicator>(),
-                                                eq(String::class.java))
-        ).thenReturn(ResponseEntity("content",HttpStatus.ACCEPTED))
+  @Test
+  fun shouldHitProperUrlWhen_no_placeHolderProvided() {
 
-        //ACTION
-        httpOutput.output(dummyRepo)
+    val httpOutput = HttpOutput("http://localhost/YYYY-MM", mockRestTemplate, LocalDate.of(2019, 1, 25))
 
-        verify(mockRestTemplate,times(1)).postForEntity(eq("http://localhost/2019-01"),
-                any<OutputIndicator>(),
-                eq(String::class.java)
-                )
-    }
+    `when`(mockRestTemplate.postForEntity(any<String>(),
+        any<OutputIndicator>(),
+        eq(String::class.java))
+    ).thenReturn(ResponseEntity("content", HttpStatus.ACCEPTED))
 
-    @Test
-    fun shouldHitProperUrlWhen_no_placeHolderProvided() {
+    //ACTION
+    httpOutput.output(dummyRepo)
 
-        val httpOutput = HttpOutput("http://localhost/YYYY-MM", mockRestTemplate, LocalDate.of(2019, 1, 25))
-
-        whenever(mockRestTemplate.postForEntity(any<String>(),
-                any<OutputIndicator>(),
-                eq(String::class.java))
-        ).thenReturn(ResponseEntity("content",HttpStatus.ACCEPTED))
-
-        //ACTION
-        httpOutput.output(dummyRepo)
-
-        verify(mockRestTemplate,times(1)).postForEntity(eq("http://localhost/YYYY-MM"),
-                any<OutputIndicator>(),
-                eq(String::class.java)
-        )
-    }
+    verify(mockRestTemplate, times(1)).postForEntity(eq("http://localhost/YYYY-MM"),
+        any<OutputIndicator>(),
+        eq(String::class.java)
+    )
+  }
 
 }
 
