@@ -3,7 +3,7 @@ package com.societegenerale.githubcrawler.parsers
 import com.societegenerale.githubcrawler.IndicatorDefinition
 import com.societegenerale.githubcrawler.parsers.SimpleFilePathParser.Companion.FILE_PATH_INFO
 import org.assertj.core.api.Assertions.assertThat
-import org.junit.Test
+import org.junit.jupiter.api.Test
 
 class SimpleFilePathParserTest {
 
