@@ -291,7 +291,7 @@ internal class GitLabResponseDecoder : Decoder {
 
             log.debug("\t ... as a " + type.typeName)
 
-            val jacksonConverter = MappingJackson2HttpMessageConverter(ObjectMapper().registerModule(KotlinModule()))
+            val jacksonConverter = MappingJackson2HttpMessageConverter(ObjectMapper().registerModule(KotlinModule.Builder().build()))
             val objectFactory = { HttpMessageConverters(jacksonConverter) }
             return ResponseEntityDecoder(SpringDecoder(objectFactory)).decode(response, type)
 
@@ -306,13 +306,13 @@ internal class GitLabResponseDecoder : Decoder {
         try {
             return repoConfigMapper.readValue(responseAsString, RepositoryConfig::class.java)
         } catch (e: IOException) {
-            throw Repository.RepoConfigException("unable to parse config for repo - content : \"$responseAsString\"", e)
+            throw Repository.RepoConfigException(HttpStatus.BAD_REQUEST,"unable to parse config for repo - content : \"$responseAsString\"", e)
         }
     }
 
-    class NoFileFoundFeignException(message: String) : FeignException(message)
+    class NoFileFoundFeignException(message: String) : FeignException(HttpStatus.NOT_FOUND.value(),message)
 
-    class GitLabException(message: String) : FeignException(message)
+    class GitLabException(message: String) : FeignException(HttpStatus.INTERNAL_SERVER_ERROR.value(),message)
 
 }
 
