@@ -1,25 +1,23 @@
 package com.societegenerale.githubcrawler;
 
+import static org.assertj.core.api.Assertions.assertThat;
+
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
 import com.fasterxml.jackson.module.kotlin.KotlinModule;
-
-import org.junit.Test;
-import org.springframework.util.StreamUtils;
-
 import java.io.IOException;
 import java.io.InputStream;
 import java.nio.charset.Charset;
+import org.junit.jupiter.api.Test;
+import org.springframework.util.StreamUtils;
 
-import static org.assertj.core.api.Assertions.assertThat;
 
-
-public class ConfigParserTest {
+class ConfigParserTest {
 
     ObjectMapper mapper = new ObjectMapper(new YAMLFactory());
 
     @Test
-    public void canParseSimpleYamlConfig() throws IOException {
+    void canParseSimpleYamlConfig() throws IOException {
 
         String configToParse = "excluded: true";
 
@@ -29,7 +27,7 @@ public class ConfigParserTest {
     }
 
     @Test
-    public void canParseMoreComplexYamlConfig() throws IOException {
+    void canParseMoreComplexYamlConfig() throws IOException {
 
         InputStream is = getClass().getClassLoader().getResourceAsStream("sampleRepoConfig.yaml");
         String configToParse = StreamUtils.copyToString(is, Charset.forName("UTF-8"));
@@ -48,7 +46,7 @@ public class ConfigParserTest {
     }
 
     @Test
-    public void canParseSimpleJsonConfig() throws IOException {
+    void canParseSimpleJsonConfig() throws IOException {
 
         String configToParse = "{\"excluded\": true}";
 
