@@ -1,21 +1,20 @@
 package com.societegenerale.githubcrawler.parsers;
 
-import com.societegenerale.githubcrawler.IndicatorDefinition;
-import org.apache.commons.io.FileUtils;
-import org.apache.commons.lang3.StringUtils;
-import org.junit.Before;
-import org.junit.Test;
-import org.springframework.util.ResourceUtils;
-
-import java.io.IOException;
-import java.util.HashMap;
-import java.util.Map;
-
 import static com.societegenerale.githubcrawler.parsers.PomXmlParserForDependencyVersion.ARTIFACT_ID;
 import static com.societegenerale.githubcrawler.parsers.PomXmlParserForDependencyVersion.FIND_DEPENDENCY_VERSION_IN_XML_METHOD;
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class PomXmlParserForDependencyVersionTest {
+import com.societegenerale.githubcrawler.IndicatorDefinition;
+import java.io.IOException;
+import java.util.HashMap;
+import java.util.Map;
+import org.apache.commons.io.FileUtils;
+import org.apache.commons.lang3.StringUtils;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.springframework.util.ResourceUtils;
+
+class PomXmlParserForDependencyVersionTest {
 
     final String indicatorName = "someIndicatorName";
     String pomXmlSnippet;
@@ -23,7 +22,7 @@ public class PomXmlParserForDependencyVersionTest {
     IndicatorDefinition pomXmlDependencyVersion = new IndicatorDefinition(indicatorName,FIND_DEPENDENCY_VERSION_IN_XML_METHOD,params);
     PomXmlParserForDependencyVersion fileContentParser = new PomXmlParserForDependencyVersion();
 
-    @Before
+    @BeforeEach
     public void setup() throws IOException {
 
         pomXmlSnippet =  FileUtils.readFileToString(ResourceUtils.getFile("classpath:sample_pom.xml"),"UTF-8");
@@ -31,35 +30,35 @@ public class PomXmlParserForDependencyVersionTest {
     }
 
     @Test
-    public void canFindAdependencyVersionInPomXml() {
+    void canFindAdependencyVersionInPomXml() {
 
         Map<String, String> indicatorsFound = parseSamplePomXmlForArtifactId("spring-boot-starter-parent");
 
-        assertThat(indicatorsFound.get(indicatorName)).isEqualTo("1.5.9.RELEASE");
+        assertThat(indicatorsFound).containsEntry(indicatorName,"1.5.9.RELEASE");
     }
 
     @Test
-    public void shouldReport_notFound_WhenArtifactIsNotFound() {
+    void shouldReport_notFound_WhenArtifactIsNotFound() {
 
         Map<String, String> indicatorsFound = parseSamplePomXmlForArtifactId("spring-not-found-parent");
 
-        assertThat(indicatorsFound.get(indicatorName)).isEqualTo("not found");
+        assertThat(indicatorsFound).containsEntry(indicatorName,"not found");
     }
 
     @Test
-    public void shouldReport_versionNotFound_WhenArtifactIsFoundButNoeTheVersion() throws Exception {
+    void shouldReport_versionNotFound_WhenArtifactIsFoundButNoeTheVersion() throws Exception {
 
         Map<String, String> indicatorsFound = parseSamplePomXmlForArtifactId("weird-artifact-with-version-declared-before");
 
-        assertThat(indicatorsFound.get(indicatorName)).isEqualTo("artifact found, but not the version");
+        assertThat(indicatorsFound).containsEntry(indicatorName,"artifact found, but not the version");
     }
 
     @Test
-    public void shouldReportVersionEvenWhenDefinedAsProperty() throws Exception {
+    void shouldReportVersionEvenWhenDefinedAsProperty() throws Exception {
 
         Map<String, String> indicatorsFound = parseSamplePomXmlForArtifactId("jackson-dataformat-yaml");
 
-        assertThat(indicatorsFound.get(indicatorName)).isEqualTo("2.9.2");
+        assertThat(indicatorsFound).containsEntry(indicatorName,"2.9.2");
     }
 
     private Map<String, String> parseSamplePomXmlForArtifactId(String artifactId) {
